# backend-saldo/src/main/java/org/farg/banking/model/Saldo.java
package org.farg.banking.model;

public class Saldo {
	
	public String moeda;
    public Double valor;

    public Saldo() {
    }

    public Saldo(String moeda, Double valor) {
        this.moeda = moeda;
        this.valor = valor;
    }
    
    
}